feat: refactor ZMQ connection handling and introduce IBridgeMessage interface
