+++ apps/vendor-ctp/src/index.ts
@@ -1,2 +1,4 @@
-import './ctp-monitor';
-import './exchange';
+import './ctp-service';
+if (!process.env.RUN_CTP_ONLY) {
+  import('./exchange');
+}
