+++ apps/vendor-ctp/src/interfaces.ts
@@ -1,0 +1,14 @@
+export interface IBridgeMessage<Req, Rep> {
+  request_id: number;
+  req?: {
+    method: string;
+    params: Req; // in JSON
+  };
+  res?: {
+    error_code: number;
+    error_message: string;
+    event: string;
+    value?: Rep; // in JSON
+    is_last: boolean;
+  };
+}
